+++ tests/test_hookman_generator.py
@@ -220,8 +220,17 @@ def test_generate_plugin_package(
     import sys
 
     shared_lib_name = f"{plugin_id}.dll" if sys.platform == "win32" else f"lib{plugin_id}.so"
-    shared_lib_path = artifacts_dir / shared_lib_name
-    shared_lib_path.write_text("")
+    if sys.platform == "win32":
+        shared_libs = [shared_lib_name]
+    else:
+        # On Linux, shared libraries often have a version suffix.
+        # For example: `libacme.so.2.3`
+        shared_dep_name = f"{plugin_id}.so.2.3"
+        shared_libs = [shared_lib_name, shared_dep_name]
+
+    for lib_name in shared_libs:
+        shared_lib_path = artifacts_dir / lib_name
+        shared_lib_path.touch()
 
     hg.generate_plugin_package(
         package_name="acme",
@@ -253,7 +262,8 @@ def test_generate_plugin_package(
 
     assert "assets/plugin.yaml" in list_of_files
     assert "assets/README.md" in list_of_files
-    assert f"artifacts/{shared_lib_name}" in list_of_files
+    for lib_name in shared_libs:
+        assert f"artifacts/{lib_name}" in list_of_files
 
     with plugin_file_zip.open("assets/plugin.yaml", "r") as f:
         contents = f.read().decode("utf-8")